pyproject.toml
[tool.poetry]
name = "tg_bot"
version = "0.1.0"
description = "A Telegram bot written in Python"
authors = ["Sebastiano Barezzi <barezzisebastiano@gmail.com>"]
# el0xren <el0xren@gmail.com>

[tool.poetry.dependencies]
python = "^3.9"  # Updated to 3.9+ for PTB v22.0
cowsay = "^6.1"
GitPython = "^3.1.14"
paramiko = "^2.7.2"
PyGithub = "^1.54.1"
python-telegram-bot = "^22.0.0"  # Updated to latest
python-dotenv = "^0.15.0"
requests = "^2.31.1"
speedtest-cli = "^2.1.2"
twrpdtgen = "^1.2.5"
natsort = "^7.1.1"
beautifulsoup4 = "^4.12.2"
ujson = "^5.8.0"
aiohttp = ">=3.8,<4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"